+++ lib/src/service/service_i.dart
@@ -17,19 +17,34 @@ abstract interface class ServiceI {
   /// Creates a new [ServiceI] instance.
   const ServiceI(this.address, this.client, this.headers);
 
+  /// Perfors a {GET} method http action to the given [endpoint].
+  ///
+  ///
+  /// [endpoint] path relative to the current handler server [address] where the action will be performed to.
+  ///
+  /// [auth] (optional) custom authentication value token.
+  ///
+  /// [headers] (optional) headers to be sent at [endpoint] request level.
+  Future<ResponseControllerI> get(
+    String endpoint, {
+    String? auth,
+    Headers? headers,
+  });
+
   /// Performs a {POST} method action via network connection to the specified {endpoint}  at the service [address].
   ///
   ///
   /// [T] type of the object model to be sent in the request body.
   ///
-  /// [endpoint] last segment of the path from the common used endpoints.
+  ///
+  /// [endpoint] path relative to the current handler server [address] where the action will be performed to.
   ///
   /// [body] the request data body.
   ///
   /// [auth] (optional) custom authentication string token.
   ///
   /// [headers] (optional) headers to be sent at [endpoint] request level.
-  Future<ResponseController> post<T extends EncodableI>(
+  Future<ResponseControllerI> post<T extends EncodableI>(
     String endpoint,
     T body, {
     String? auth,
@@ -41,15 +56,16 @@ abstract interface class ServiceI {
   ///
   /// [T] type of the object model list to be sent in the request body.
   ///
-  /// [endpoint] last segment of the path from the common used endpoints.
+  ///
+  /// [endpoint] path relative to the current handler server [address] where the action will be performed to.
   ///
   /// [body] the request data body.
   ///
   /// [auth] (optional) custom authentication string token.
   ///
   /// [headers] (optional) headers to be sent at [endpoint] request level.
-  Future<ResponseController> postList<T extends EncodableI>(
-    String act,
+  Future<ResponseControllerI> postList<T extends EncodableI>(
+    String endpoint,
     List<T> request, {
     String? auth,
     Headers? headers,